Move SplitterUI under laf package
